SSA CFG: move ValueId::str into its own SSACFGTypes.cpp
